chore: re-organise volatile files.
