--- crates/amaru-ledger/src/state/volatile_db.rs
@@ -1,753 +1,0 @@
-// Copyright 2024 PRAGMA
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::{
-    collections::{BTreeMap, BTreeSet, VecDeque},
-    mem,
-};
-
-use amaru_kernel::{
-    Anchor, Ballot, BallotId, CertificatePointer, ComparableProposalId, DRep, DRepRegistration, Epoch, Lovelace,
-    MemoizedTransactionOutput, Point, PoolId, PoolParams, Proposal, ProposalPointer, ProtocolParameters,
-    StakeCredential, Tip, TransactionInput,
-};
-
-use super::{
-    diff_bind::{Bind, DiffBind, Empty},
-    diff_epoch_reg::DiffEpochReg,
-    diff_set::DiffSet,
-};
-use crate::{
-    governance::ratification::ProposalsRootsRc,
-    state::{
-        diff_bind::{MergeError, Resettable},
-        diff_epoch_reg::Registrations,
-    },
-    store::{
-        self, ReadStore, StoreError,
-        columns::{pools::Row as Pool, *},
-    },
-};
-
-// VolatileDB
-// ----------------------------------------------------------------------------
-
-#[derive(Default)]
-pub struct VolatileDB {
-    sequence: VecDeque<AnchoredVolatileState>,
-}
-
-impl VolatileDB {
-    pub fn is_empty(&self) -> bool {
-        self.sequence.is_empty()
-    }
-
-    pub fn len(&self) -> usize {
-        self.sequence.len()
-    }
-
-    pub fn view_back(&self) -> Option<&AnchoredVolatileState> {
-        self.sequence.back()
-    }
-
-    pub fn view_front(&self) -> Option<&AnchoredVolatileState> {
-        self.sequence.front()
-    }
-
-    pub fn resolve_input(&self, input: &TransactionInput) -> Option<&MemoizedTransactionOutput> {
-        for state in self.sequence.iter().rev() {
-            if state.state.utxo.consumed.contains(input) {
-                return None;
-            }
-            if let Some(output) = state.state.utxo.produced.get(input) {
-                return Some(output);
-            }
-        }
-        None
-    }
-
-    pub fn has_consumed_input(&self, input: &TransactionInput) -> bool {
-        self.sequence.iter().any(|state| state.state.utxo.consumed.contains(input))
-    }
-
-    pub fn contains(&self, point: &Point) -> bool {
-        self.sequence.binary_search_by_key(point, |state| state.anchor.0.point()).is_ok()
-    }
-
-    pub fn pop_front(&mut self) -> Option<AnchoredVolatileState> {
-        self.sequence.pop_front()
-    }
-
-    pub fn push_back(&mut self, state: AnchoredVolatileState) {
-        self.sequence.push_back(state);
-    }
-
-    pub fn rollback_to<'a>(&mut self, point: &'a Point) -> Result<(), &'a Point> {
-        let target_slot = point.slot_or_default();
-
-        // Check if the target point is beyond the sequence
-        // In this case we simply return Ok since it this would not change the volatile state.
-        if let Some(last) = self.sequence.back()
-            && last.anchor.0.slot() < target_slot
-        {
-            tracing::warn!(
-                name: "rollback_to.beyond",
-                %target_slot,
-                last_slot = ?last.anchor.0.slot(),
-                "Attempting to rollback to a point beyond the last known volatile state"
-            );
-            return Ok(());
-        }
-
-        // Check if the target point is before the sequence
-        // In this case we return an error since it means rolling back the stable DB
-        if let Some(first) = self.sequence.front()
-            && target_slot < first.anchor.0.slot()
-        {
-            tracing::error!(
-                name: "rollback_to.before",
-                %target_slot,
-                first_slot = ?first.anchor.0.slot(),
-                "Attempting to rollback to a point before the first point of the volatile state"
-            );
-            return Err(point);
-        }
-
-        // Now we know the target point is within the sequence.
-        // Keep all elements with point <= target point.
-        let mut ix = 0;
-        let mut found = false;
-        for diff in self.sequence.iter() {
-            if diff.anchor.0.point() <= *point {
-                ix += 1;
-                if diff.anchor.0.point() == *point {
-                    found = true;
-                    break;
-                }
-            } else {
-                return Err(point);
-            }
-        }
-
-        if !found {
-            return Err(point);
-        }
-
-        self.sequence.truncate(ix);
-        Ok(())
-    }
-
-    pub fn clear(&mut self) {
-        self.sequence.clear();
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum ViewError {
-    #[error(
-        "unable to construct volatile view: invariant violation ({:?}) when processing accounts at anchor {:?}:\nnext_accounts: {:#?}\ncurrent_accounts: {:#?}",
-         .merge_error,
-         .anchor,
-         .next_accounts,
-         .current_accounts,
-    )]
-    AccountsMergeError {
-        anchor: Box<(Tip, PoolId)>,
-        merge_error: MergeError<StakeCredential>,
-        next_accounts:
-            Box<DiffBind<StakeCredential, (PoolId, CertificatePointer), (DRep, CertificatePointer), Lovelace>>,
-        current_accounts:
-            Box<DiffBind<StakeCredential, (PoolId, CertificatePointer), (DRep, CertificatePointer), Lovelace>>,
-    },
-}
-
-impl ViewError {
-    pub fn accounts_merge_error(
-        anchored: &AnchoredVolatileState,
-        merge_error: MergeError<&StakeCredential>,
-        current_accounts: DiffBind<
-            &StakeCredential,
-            &(PoolId, CertificatePointer),
-            &(DRep, CertificatePointer),
-            &Lovelace,
-        >,
-    ) -> Self {
-        Self::AccountsMergeError {
-            merge_error: merge_error.to_owned(),
-            anchor: Box::new(anchored.anchor),
-            next_accounts: Box::new(anchored.state.accounts.clone()),
-            current_accounts: Box::new(current_accounts.to_owned()),
-        }
-    }
-}
-
-// VolatileState
-// ----------------------------------------------------------------------------
-
-/// Resulting state change coming from processing a block.
-#[derive(Debug, Default)]
-pub struct VolatileState {
-    pub utxo: DiffSet<TransactionInput, MemoizedTransactionOutput>,
-    pub pools: DiffEpochReg<PoolId, (PoolParams, CertificatePointer)>,
-    pub accounts: DiffBind<StakeCredential, (PoolId, CertificatePointer), (DRep, CertificatePointer), Lovelace>,
-    pub dreps: DiffBind<StakeCredential, Anchor, Empty, DRepRegistration>,
-    pub dreps_deregistrations: BTreeMap<StakeCredential, CertificatePointer>,
-    pub committee: DiffBind<StakeCredential, StakeCredential, Empty, Empty>,
-    pub withdrawals: BTreeSet<StakeCredential>,
-    pub proposals: BTreeMap<ComparableProposalId, (Proposal, ProposalPointer)>,
-    pub votes: DiffSet<BallotId, Ballot>,
-    pub fees: Lovelace,
-}
-
-pub struct AnchoredVolatileState {
-    pub anchor: (Tip, PoolId),
-    pub state: VolatileState,
-}
-
-impl VolatileState {
-    pub fn anchor(self, tip: Tip, issuer: PoolId) -> AnchoredVolatileState {
-        AnchoredVolatileState { anchor: (tip, issuer), state: self }
-    }
-
-    pub fn resolve_input(&self, input: &TransactionInput) -> Option<&MemoizedTransactionOutput> {
-        self.utxo.produced.get(input)
-    }
-
-    pub fn has_consumed_input(&self, input: &TransactionInput) -> bool {
-        self.utxo.consumed.contains(input)
-    }
-}
-
-// VolatileView
-// ----------------------------------------------------------------------------
-
-/// An aggregate of multiple VolatileState, useful at epoch boundaries or for building context.
-#[derive(Debug)]
-pub struct VolatileView<'volatile, 'store, DB: ReadStore> {
-    epoch: Epoch,
-    db: &'store DB,
-    pools: Option<DiffEpochReg<PoolId, &'volatile (PoolParams, CertificatePointer)>>,
-    proposals: BTreeMap<&'volatile ComparableProposalId, &'volatile (Proposal, ProposalPointer)>,
-    accounts: Option<AccountVolatileView<'volatile>>,
-}
-
-/// A simplified 'DiffBind' for accounts, specialized to just the stake credentials.
-#[derive(Debug)]
-struct AccountVolatileView<'volatile> {
-    registered: BTreeSet<&'volatile StakeCredential>,
-    unregistered: BTreeSet<&'volatile StakeCredential>,
-}
-
-/// An internal iterator that proxies the stable store's `iter_pools`, but taking into account any
-/// pending volatile update.
-///
-/// NOTE: About 'IterPools'
-///
-/// This iterator may look complicated, but it exists for two reasons:
-///
-/// 1. It allows to stick with iterators; which means that the compiler and execution can be
-///    optimised towards that. There's no overhead resulting from allocating a large vector of 3000
-///    pools. We can intead rely on streaming all the way to construct the updates.
-///
-/// 2. Until we have hit the stable store, we cannot know whether a registration is actually a new
-///    registration, or if it's a re-registration. That's because there's no mechanism to 'update'
-///    a pool really, they just re-register. Yet, we don't want to be inspecting each registration
-///    independently with a db call because that could be very dramatic.
-///
-///    Since we can process those updates in any order; we can go ahead and first treat all the
-///    database pools, an then, continue the iterator with what's left in the pending state if any.
-///    Yet, because of Rust borrowing and ownership model, we cannot just do that on top of the db
-///    iterator alone; we must introduce an extra wrapper that will own all the required data and
-///    take care of the chaining.
-///
-/// Importantly, the last points means that there's no guaranteed order on this iterator. Pools
-/// shall be considered unordered by consumers of this iterator.
-struct IterPools<'volatile, DBIter: Iterator<Item = (PoolId, Pool)>> {
-    epoch: Epoch,
-    db_iterator: DBIter,
-    registrations: BTreeMap<PoolId, Registrations<&'volatile (PoolParams, CertificatePointer)>>,
-    retirements: BTreeMap<PoolId, Epoch>,
-}
-
-impl<'volatile, DBIter: Iterator<Item = (PoolId, Pool)>> Iterator for IterPools<'volatile, DBIter> {
-    type Item = (PoolId, Pool);
-
-    // TODO: reduce logic duplication?
-    //
-    // - The following code 'patches' the immutable db state with what's transient in the
-    //   volatile.
-    //
-    // - Fundamentally, it duplicates the logic of:
-    //   - state::volatile_db::add_pools
-    //   - store::columns::pools::extend
-    //   - rocksdb::ledger::columns::pools::{add, remove}
-    //
-    // - However, it doesn't duplicate things in a way that's trivial to unify. But that's
-    //   probably something we may want to look into? Perhaps as one of the design goal for a
-    //   future ledger store.
-    //
-    // TODO: annoying clones
-    //
-    // This also contains a few annoying clones which could likely be avoided or deferred by having
-    // the iterator works over a `&Pool`.
-    fn next(&mut self) -> Option<Self::Item> {
-        // First, we patch stable pools with any pending update
-        if let Some((pool_id, mut pool)) = self.db_iterator.next() {
-            // Pool is already registered, and has some updates.
-            if let Some(update) = self.registrations.remove(&pool_id) {
-                let mut future_params =
-                    update.into_iter().map(|(pool_params, _)| (Some(pool_params.clone()), self.epoch + 1)).collect();
-                pool.future_params.append(&mut future_params);
-            }
-
-            // Pool has announced its retirement.
-            if let Some(retirement_epoch) = self.retirements.remove(&pool_id) {
-                pool.future_params.append(&mut vec![(None, retirement_epoch)])
-            }
-
-            return Some((pool_id, pool));
-        }
-
-        // Then, we must add any pool that only appears in the volatile
-        if let Some((pool_id, registrations)) = self.registrations.pop_first() {
-            let (registration, re_registration) = registrations.into_inner();
-
-            let mut pool = Pool::new(registration.1, registration.0.clone());
-            if let Some(re_registration) = re_registration {
-                pool.future_params = vec![(Some(re_registration.0.clone()), self.epoch + 1)]
-            }
-
-            return Some((pool_id, pool));
-        }
-
-        None
-    }
-}
-
-/// Similar to [`IterPools`], but for accounts; making sure we acknowledge deregistrations while
-/// continuing with new fresh registrations.
-struct IterAccounts<'volatile, DBIter: Iterator<Item = (StakeCredential, accounts::Row)>> {
-    db_iterator: DBIter,
-    registrations: BTreeSet<&'volatile StakeCredential>,
-    deregistrations: BTreeSet<&'volatile StakeCredential>,
-}
-
-impl<'volatile, DBIter: Iterator<Item = (StakeCredential, accounts::Row)>> Iterator
-    for IterAccounts<'volatile, DBIter>
-{
-    type Item = StakeCredential;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some((account, _)) = self.db_iterator.next()
-            && !self.deregistrations.contains(&account)
-        {
-            return Some(account);
-        }
-
-        if let Some(account) = self.registrations.pop_first() {
-            return Some(account.clone());
-        }
-
-        None
-    }
-}
-
-impl<'volatile, 'db, DB: ReadStore> VolatileView<'volatile, 'db, DB> {
-    /// Obtain a view of the database, which acts as a proxy 'ReadStore' augmented with the latest
-    /// volatile updates, if any. This is used in context where one needs the true latest view of
-    /// the ledger; for example at the epoch boundary.
-    pub fn new(
-        // TODO: Derive epoch instead of taking extra arg
-        //
-        // Currently passing this argument for simplicity, but that's a door open to
-        // inconsistencies. In principle we should be able to derive the epoch from either the
-        // stable db or self; since there can be only epoch in the context where this function is
-        // called. It's even an invariant violation if not...
-        epoch: Epoch,
-        volatile: &'volatile VolatileDB,
-        stable: &'db DB,
-    ) -> Result<VolatileView<'volatile, 'db, DB>, ViewError> {
-        let mut pools = DiffEpochReg::default();
-        let mut proposals = BTreeMap::new();
-        let mut accounts = DiffBind::default();
-
-        for anchored in volatile.sequence.iter() {
-            if let Err(merge_error) = accounts.evolve(anchored.state.accounts.into_borrowed()) {
-                return Err(ViewError::accounts_merge_error(anchored, merge_error, accounts));
-            }
-
-            pools.evolve(anchored.state.pools.into_borrowed());
-
-            for (k, v) in anchored.state.proposals.iter() {
-                proposals.insert(k, v);
-            }
-        }
-
-        let accounts = AccountVolatileView {
-            registered: accounts.registered.into_keys().collect(),
-            unregistered: accounts.unregistered,
-        };
-
-        Ok(Self { db: stable, epoch, accounts: Some(accounts), pools: Some(pools), proposals })
-    }
-
-    /// Provides an iterator for pools on top of the stable store, but adding any pending updates
-    /// from the aggregated volatile state.
-    ///
-    /// IMPORTANT: Yields pools in no particular order.
-    pub fn iter_pools(&mut self) -> Result<impl Iterator<Item = (PoolId, Pool)>, StoreError> {
-        match mem::take(&mut self.pools) {
-            None => {
-                // Just being careful here. There's no reason to ever call this twice; but if it
-                // ever happens, this line might save us from hours of debugging.
-                unreachable!(".iter_pools() called twice on the same VolatileView! Don't do that.")
-            }
-            Some(mut pools) => Ok(IterPools {
-                epoch: self.epoch,
-                db_iterator: self.db.iter_pools()?,
-                registrations: mem::take(&mut pools.registered),
-                retirements: mem::take(&mut pools.unregistered),
-            }),
-        }
-    }
-
-    /// Provides an iterator for proposals on top of the stable store, but adding any pending updates
-    /// from the aggregated volatile state.
-    ///
-    /// IMPORTANT: Yields proposals in no particular order.
-    pub fn iter_proposals(&self) -> Result<impl Iterator<Item = (ComparableProposalId, proposals::Row)>, StoreError> {
-        Ok(self
-            .db
-            .iter_proposals()?
-            .chain(add_proposals(self.proposals.iter().map(|(k, v)| ((*k).clone(), (*v).clone())), self.epoch)))
-    }
-
-    /// Provides an iterator for accounts on top of the stable store, also applying any pending
-    /// registration or deregistration from the aggregated volatile state.
-    ///
-    /// IMPORTANT: Yields accounts in no particular order.
-    pub fn iter_accounts(&mut self) -> Result<impl Iterator<Item = StakeCredential>, StoreError> {
-        match mem::take(&mut self.accounts) {
-            None => {
-                // Just being careful here. There's no reason to ever call this twice; but if it
-                // ever happens, this line might save us from hours of debugging.
-                unreachable!(".iter_accounts() called twice on the same VolatileView! Don't do that.")
-            }
-            Some(mut accounts) => Ok(IterAccounts {
-                db_iterator: self.db.iter_accounts()?,
-                registrations: mem::take(&mut accounts.registered),
-                deregistrations: mem::take(&mut accounts.unregistered),
-            }),
-        }
-    }
-
-    /// A view on the proposal roots; this doesn't really require any volatile update but is
-    /// conveniently made available from the underlying store; to avoid having to pass both a
-    /// volatile view and a stable store around every function.
-    pub fn proposals_roots(&self) -> Result<ProposalsRootsRc, StoreError> {
-        Ok(ProposalsRootsRc::from(self.db.proposals_roots()?))
-    }
-}
-
-// StoreUpdate
-// ----------------------------------------------------------------------------
-
-pub struct StoreUpdate<W, A, R> {
-    pub point: Point,
-    pub issuer: PoolId,
-    pub fees: Lovelace,
-    pub withdrawals: W,
-    pub add: A,
-    pub remove: R,
-}
-
-impl AnchoredVolatileState {
-    #[expect(clippy::type_complexity)]
-    pub fn into_store_update(
-        self,
-        epoch: Epoch,
-        protocol_parameters: &ProtocolParameters,
-    ) -> StoreUpdate<
-        impl Iterator<Item = accounts::Key>,
-        store::Columns<
-            impl Iterator<Item = (utxo::Key, utxo::Value)>,
-            impl Iterator<Item = pools::Value>,
-            impl Iterator<Item = (accounts::Key, accounts::Value)>,
-            impl Iterator<Item = (dreps::Key, dreps::Value)>,
-            impl Iterator<Item = (cc_members::Key, cc_members::Value)>,
-            impl Iterator<Item = (proposals::Key, proposals::Value)>,
-            impl Iterator<Item = (votes::Key, votes::Value)>,
-        >,
-        store::Columns<
-            impl Iterator<Item = utxo::Key>,
-            impl Iterator<Item = (pools::Key, Epoch)>,
-            impl Iterator<Item = accounts::Key>,
-            impl Iterator<Item = (dreps::Key, CertificatePointer)>,
-            impl Iterator<Item = cc_members::Key>,
-            impl Iterator<Item = ()>,
-            impl Iterator<Item = ()>,
-        >,
-    > {
-        let gov_action_lifetime = protocol_parameters.gov_action_lifetime;
-
-        StoreUpdate {
-            point: self.anchor.0.point(),
-            issuer: self.anchor.1,
-            fees: self.state.fees,
-            withdrawals: self.state.withdrawals.into_iter(),
-            add: store::Columns {
-                utxo: self.state.utxo.produced.into_iter(),
-                pools: add_pools(self.state.pools.registered.into_iter(), epoch),
-                accounts: add_accounts(self.state.accounts.registered.into_iter()),
-                dreps: add_dreps(self.state.dreps.registered.into_iter()),
-                cc_members: add_committee(self.state.committee.registered.into_iter()),
-                proposals: add_proposals(self.state.proposals.into_iter(), epoch + gov_action_lifetime),
-                votes: self.state.votes.produced.into_iter(),
-            },
-            remove: store::Columns {
-                utxo: self.state.utxo.consumed.into_iter(),
-                pools: self.state.pools.unregistered.into_iter(),
-                accounts: self.state.accounts.unregistered.into_iter(),
-                dreps: remove_dreps(self.state.dreps.unregistered.into_iter(), self.state.dreps_deregistrations),
-                cc_members: self.state.committee.unregistered.into_iter(),
-                proposals: std::iter::empty(),
-                votes: {
-                    debug_assert!(self.state.votes.consumed.is_empty());
-                    std::iter::empty()
-                },
-            },
-        }
-    }
-}
-
-// -------------------------------------------------------------------- Pools
-// --------------------------------------------------------------------------
-
-fn add_pools(
-    iterator: impl Iterator<Item = (PoolId, Registrations<(PoolParams, CertificatePointer)>)>,
-    epoch: Epoch,
-) -> impl Iterator<Item = pools::Value> {
-    iterator.flat_map(move |(_, registrations)| {
-        registrations
-            .into_iter()
-            // NOTE/TODO: Re-registrations (a.k.a pool params updates) are always
-            // happening on the following epoch. We do not explicitly store epochs
-            // for registrations in the DiffEpochReg (which may be an arguable
-            // choice?) so we have to artificially set it here. Note that for
-            // registrations (when there's no existing entry), the epoch is wrong
-            // but it is fully ignored. It's slightly ugly, but we cannot know if
-            // an entry exists without querying the stable store -- and frankly, we
-            // don't _have to_.
-            .map(|registration| (registration.0, registration.1, epoch + 1))
-            .collect::<Vec<_>>()
-    })
-}
-
-// ----------------------------------------------------------------- Accounts
-// --------------------------------------------------------------------------
-
-fn add_accounts(
-    iterator: impl Iterator<
-        Item = (StakeCredential, Bind<(PoolId, CertificatePointer), (DRep, CertificatePointer), Lovelace>),
-    >,
-) -> impl Iterator<Item = (accounts::Key, accounts::Value)> {
-    iterator
-        .map(|(credential, Bind { left: pool, right: drep, value: deposit })| (credential, (pool, drep, deposit, 0)))
-}
-
-// -------------------------------------------------------------------- DReps
-// --------------------------------------------------------------------------
-
-fn add_dreps(
-    iterator: impl Iterator<Item = (StakeCredential, Bind<Anchor, Empty, DRepRegistration>)>,
-) -> impl Iterator<Item = (dreps::Key, dreps::Value)> {
-    iterator.map(move |(credential, Bind { left: anchor, right: _, value: registration }): (_, Bind<_, Empty, _>)| {
-        (credential, (anchor, registration))
-    })
-}
-
-fn remove_dreps(
-    iterator: impl Iterator<Item = StakeCredential>,
-    mut deregistrations: BTreeMap<StakeCredential, CertificatePointer>,
-) -> impl Iterator<Item = (dreps::Key, CertificatePointer)> {
-    iterator.map(move |credential| {
-        #[expect(clippy::expect_used)]
-        let pointer =
-            deregistrations.remove(&credential).expect("every 'unregistered' drep must have a matching deregistration");
-
-        (credential, pointer)
-    })
-}
-
-// --------------------------------------------------------------- CC Members
-// --------------------------------------------------------------------------
-
-fn add_committee(
-    iterator: impl Iterator<Item = (StakeCredential, Bind<StakeCredential, Empty, Empty>)>,
-) -> impl Iterator<Item = (cc_members::Key, cc_members::Value)> {
-    iterator.map(|(credential, Bind { left: hot_credential, right: _, value: _ })| {
-        (credential, (hot_credential, Resettable::Unchanged))
-    })
-}
-
-// ---------------------------------------------------------------- Proposals
-// --------------------------------------------------------------------------
-
-fn add_proposals(
-    iterator: impl Iterator<Item = (ComparableProposalId, (Proposal, ProposalPointer))>,
-    expiration: Epoch,
-) -> impl Iterator<Item = (proposals::Key, proposals::Value)> {
-    iterator.map(move |(proposal_id, (proposal, proposed_in))| {
-        (proposal_id, proposals::Value { proposed_in, valid_until: expiration, proposal })
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use amaru_kernel::{BlockHeight, Hash, Point, Slot};
-
-    use super::*;
-
-    #[test]
-    fn test_rollback_to_point_before_sequence_fails() {
-        // Create a VolatileDB with three states at slots 10, 20, 30
-        let mut db = create_volatile_db();
-
-        // Rollback to slot 5 (before the first element at slot 10)
-        // This represents rolling back to a point in the stable DB
-        let rollback_point = Point::Specific(Slot::from(5), Hash::new([0u8; 32]));
-
-        let result = db.rollback_to(&rollback_point);
-
-        // This should fail
-        // (rolling back to a point inside the stable DB is not allowed)
-        assert!(result.is_err());
-        assert_eq!(db.len(), 3, "All elements should be retained");
-    }
-
-    #[test]
-    fn test_rollback_to_exact_last_element_should_succeed() {
-        // Create a VolatileDB with three states at slots 10, 20, 30
-        let mut db = create_volatile_db();
-
-        // Rollback to slot 30 (the last element)
-        let rollback_point = Point::Specific(Slot::from(30), Hash::new([0u8; 32]));
-
-        // This should succeed, keeping all 3 elements
-        let result = db.rollback_to(&rollback_point);
-
-        assert!(result.is_ok(), "Rolling back to the exact slot of the last element should succeed");
-        assert_eq!(db.len(), 3, "All elements should be retained");
-    }
-
-    #[test]
-    fn test_rollback_to_middle_element_succeeds() {
-        // Create a VolatileDB with three states at slots 10, 20, 30
-        let mut db = create_volatile_db();
-
-        // Rollback to slot 20 (middle element)
-        let rollback_point = Point::Specific(Slot::from(20), Hash::new([0u8; 32]));
-
-        let result = db.rollback_to(&rollback_point);
-
-        // This should succeed
-        assert!(result.is_ok());
-        assert_eq!(db.len(), 2, "Should keep elements at slots 10 and 20");
-    }
-
-    #[test]
-    fn test_rollback_to_point_after_sequence_succeeds() {
-        // Create a VolatileDB with three states at slots 10, 20, 30
-        let mut db = create_volatile_db();
-
-        // Try to rollback to slot 40 (after the sequence)
-        let rollback_point = Point::Specific(Slot::from(40), Hash::new([0u8; 32]));
-
-        let result = db.rollback_to(&rollback_point);
-
-        // This should succeed
-        assert!(result.is_ok(), "Rolling back to a point after the sequence should succeed");
-        assert_eq!(db.len(), 3, "All elements should be retained");
-    }
-
-    #[test]
-    fn test_rollback_to_slot_between_elements_succeeds() {
-        // Create a VolatileDB with three states at slots 10, 20, 30
-        let mut db = create_volatile_db();
-
-        // Rollback to slot 25 (between 20 and 30)
-        let rollback_point = Point::Specific(Slot::from(25), Hash::new([0u8; 32]));
-
-        let result = db.rollback_to(&rollback_point);
-
-        assert_eq!(result.unwrap_err(), &rollback_point);
-        assert_eq!(db.len(), 3, "All elements should be retained");
-    }
-
-    #[test]
-    fn test_consumed_input_is_tracked() {
-        let input = test_input(1);
-        let mut state = create_test_state(10, 1);
-        state.state.utxo.consume(input.clone());
-
-        let mut db = VolatileDB::default();
-        db.push_back(state);
-
-        assert!(db.has_consumed_input(&input));
-        assert!(db.resolve_input(&input).is_none());
-    }
-
-    #[test]
-    fn test_rollback_removes_consumed_input_from_cache() {
-        let input = test_input(1);
-        let mut db = VolatileDB::default();
-        let first = create_test_state(10, 1);
-        let first_point = first.anchor.0.point();
-        db.push_back(first);
-
-        let mut second = create_test_state(20, 2);
-        second.state.utxo.consume(input.clone());
-        db.push_back(second);
-
-        assert!(db.has_consumed_input(&input));
-
-        db.rollback_to(&first_point).unwrap();
-
-        assert!(!db.has_consumed_input(&input));
-    }
-
-    // HELPERS
-
-    fn create_test_state(slot: u64, pool_id: u8) -> AnchoredVolatileState {
-        let point = Point::Specific(Slot::from(slot), Hash::new([0u8; 32]));
-        let pool = Hash::new([pool_id; 28]);
-        let tip = Tip::new(point, BlockHeight::from(slot));
-
-        AnchoredVolatileState { anchor: (tip, pool), state: VolatileState::default() }
-    }
-
-    fn create_volatile_db() -> VolatileDB {
-        let mut db = VolatileDB::default();
-        db.push_back(create_test_state(10, 1));
-        db.push_back(create_test_state(20, 2));
-        db.push_back(create_test_state(30, 3));
-
-        assert_eq!(db.len(), 3);
-        db
-    }
-
-    fn test_input(tag: u8) -> TransactionInput {
-        TransactionInput { transaction_id: Hash::new([tag; 32]), index: 0 }
-    }
-}
